Use a non-colliding package in bundled package tests

+++ x-pack/test/fleet_api_integration/apis/epm/install_bundled.ts
@@ -45,7 +45,12 @@ export default function (providerContext: FtrProviderContext) {
       const files = await fs.readdir(BUNDLED_PACKAGES_DIR);
 
       for (const file of files) {
-        await fs.unlink(path.join(BUNDLED_PACKAGES_DIR, file));
+        const isFixtureFile = !!(await fs.readFile(path.join(BUNDLED_PACKAGE_FIXTURES_DIR, file)));
+
+        // Only remove fixture files - leave normal bundled packages in place
+        if (isFixtureFile) {
+          await fs.unlink(path.join(BUNDLED_PACKAGES_DIR, file));
+        }
       }
     } catch (error) {
       log.error('Error removing bundled packages');
@@ -63,10 +68,11 @@ export default function (providerContext: FtrProviderContext) {
 
     describe('without registry', () => {
       it('installs from bundled source via api', async () => {
-        await bundlePackage('elastic_agent-1.2.0');
+        // Need to bundle a package that doesn't conflict with those listed in `fleet_packages.json
+        await bundlePackage('nginx-1.2.1');
 
         const response = await supertest
-          .post(`/api/fleet/epm/packages/elastic_agent/1.2.0`)
+          .post(`/api/fleet/epm/packages/nginx/1.2.1`)
           .set('kbn-xsrf', 'xxxx')
           .type('application/json')
           .send({ force: true })
@@ -76,8 +82,8 @@ export default function (providerContext: FtrProviderContext) {
       });
 
       it('allows for upgrading from newer bundled source when outdated package was installed from bundled source', async () => {
-        await bundlePackage('elastic_agent-1.0.0');
-        await bundlePackage('elastic_agent-1.2.0');
+        await bundlePackage('nginx-1.1.0');
+        await bundlePackage('nginx-1.2.1');
 
         const installResponse = await supertest
           .post(`/api/fleet/epm/packages/elastic_agent/1.0.0`)
@@ -89,7 +95,7 @@ export default function (providerContext: FtrProviderContext) {
         expect(installResponse.body._meta.install_source).to.be('bundled');
 
         const updateResponse = await supertest
-          .post(`/api/fleet/epm/packages/elastic_agent/1.2.0`)
+          .post(`/api/fleet/epm/packages/nginx/1.2.1`)
           .set('kbn-xsrf', 'xxxx')
           .type('application/json')
           .send({ force: true })
@@ -101,10 +107,10 @@ export default function (providerContext: FtrProviderContext) {
 
     describe('with registry', () => {
       it('allows for updating from registry when outdated package is installed from bundled source', async () => {
-        await bundlePackage('elastic_agent-1.2.0');
+        await bundlePackage('nginx-1.1.0');
 
         const bundledInstallResponse = await supertest
-          .post(`/api/fleet/epm/packages/elastic_agent/1.2.0`)
+          .post(`/api/fleet/epm/packages/nginx/1.1.0`)
           .set('kbn-xsrf', 'xxxx')
           .type('application/json')
           .send({ force: true })
@@ -112,8 +118,9 @@ export default function (providerContext: FtrProviderContext) {
 
         expect(bundledInstallResponse.body._meta.install_source).to.be('bundled');
 
+        // Update to one version prior to the bundled version of nginx
         const registryUpdateResponse = await supertest
-          .post(`/api/fleet/epm/packages/elastic_agent/1.3.0`)
+          .post(`/api/fleet/epm/packages/elastic_agent/1.2.0`)
           .set('kbn-xsrf', 'xxxx')
           .type('application/json')
           .send({ force: true })
